fix(delete all files): Check if file has an URL before delete from storage

+++ api/router/file.router.js
@@ -66,11 +66,7 @@ file.put('/:uid/:fileId', async (req, res) => {
                     // Rename in firestore
                     await fileRef.update({
                         [fileId]: {
-                            name: `${ file.name }`,
-                            type: file.type,
-                            date: file.date,
-                            size: file.size,
-                            url: newUrl[0]
+                            name: `${ file.name }`, type: file.type, date: file.date, size: file.size, url: newUrl[0]
                         }
                     })
                         .then(() => {
@@ -94,10 +90,7 @@ file.put('/:uid/:fileId', async (req, res) => {
             // Rename in firestore
             await fileRef.update({
                 [fileId]: {
-                    name: file.name,
-                    type: file.type,
-                    size: file.size,
-                    date: file.date,
+                    name: file.name, type: file.type, size: file.size, date: file.date,
                 }
             })
                 .then(() => {
@@ -173,15 +166,17 @@ file.post('/deleteAll', async (req, res) => {
     for (const fileId in files) {
         const file = files[fileId];
 
-        // Delete all files from storage
-        admin.storage().bucket()
-            .file(`files/${ uid }/${ file.name }`)
-            .delete()
-            .catch(error => {
-                return res.status(500).json({
-                    message: error.message
-                });
-            })
+        if (file.url) {
+            // Delete all files from storage
+            admin.storage().bucket()
+                .file(`files/${ uid }/${ file.name }`)
+                .delete()
+                .catch(error => {
+                    return res.status(500).json({
+                        message: error.message
+                    });
+                })
+        }
     }
 
     // Delete all files from firestore
@@ -192,11 +187,12 @@ file.post('/deleteAll', async (req, res) => {
             res.status(200).json({
                 message: 'All files deleted successfully'
             })
-        }).catch(error => {
-        res.status(500).json({
-            message: error.message
-        });
-    })
+        })
+        .catch(error => {
+            res.status(500).json({
+                message: error.message
+            });
+        })
 });
 
 module.exports = file;
